Change int3 to float3 in parallel approaches

+++ ElfColCnt/cuda_halfsteps/test.cuh
@@ -4,7 +4,7 @@
 
 #include "ElfColCnt.cuh"
 
-int test_count(int3 *vector, int size, int iters){
+int test_count(float3 *vector, int size, int iters){
 	struct CollisionCountPromise *promises;
 	promises = (struct CollisionCountPromise *) malloc(sizeof(struct CollisionCountPromise) * iters);
